backend/requirements.txt
# Django and REST Framework
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9

# Utilities
python-decouple>=3.8
Pillow>=10.0.0

# Development
django-extensions>=3.2.3
